Mapping Hurricane Irma in real-time
Over the coming days, we will map Irma’s path of destruction using (close to) real-time information about the online/offline activity of 401,682 IP-addresses in the affected regions (Cuba, Dominican Republic, Haiti, Florida).

We combine measurements of whether an IP address is online or offline with its geo-location and map the changes of Internet activity at an hourly rate. The results are presented in an animated map below.

Our aim is to study the impact of Irma’s passage on the ICT infrastructure of affected regions.

Of course, this approach does not provide an accurate reflection of the actual damages on the ground. However, the idea is to provide a first picture where Irma’s path resulted in actual impact on society. It can be used as the very first input to assess the geographic extent of Hurricane Irma and thereby potentially offer useful information for first responders.

The data contains

IP count
Timestamp (AST)
Longitude/Latitude

If you